docs : backend 폴더구조 번경 - controller와 routes 폴더로 기능에 따라 정리
